Add unit tests for clipboard
Defined an intermediate class used to enable the
mocking of the underlying clipboard operating
system calls.  Define CLIPBOARD_REVEAL_OS_API in order
to gain access to it.  It's not defined by default
because defining it exposes operating system headers
and this is something we want to avoid in normal usage.